+++ contracts/interfaces/IJuiceStakerActions.sol
@@ -13,11 +13,10 @@ struct StakingParam {
 }
 
 interface IJuiceStakerActions {
-
     /// @notice Deposits JUICE tokens to be used in staking. Moves `amount` of JUICE from user's balance to
     /// staking contract's balance.
     /// @param amount The deposited amount. If it exceeds user's balance, tx reverts with `InsufficientJUICE` error.
-    function deposit(uint amount) external;
+    function deposit(uint256 amount) external;
 
     /// @notice Modifies the user's token stakes.
     /// @param stakes The array of StakingParams which are processed in order.
@@ -26,17 +25,17 @@ interface IJuiceStakerActions {
     /// @notice Withdraws JUICE tokens from the staking contract. Moves `amount` of JUICE from the contract's balance to
     /// user's balance.
     /// @param amount The withdrawn amount. If it exceeds user's unstaked balance, tx reverts with `InsufficientJUICE` error.
-    function withdraw(uint amount) external;
+    function withdraw(uint256 amount) external;
 
     /// @notice Emitted on successful deposit()
     /// @param user The user who made the deposit
     /// @param amount The deposited JUICE amount
-    event JUICEDeposited(address indexed user, uint amount);
+    event JUICEDeposited(address indexed user, uint256 amount);
 
     /// @notice Emitted on successful withdraw()
     /// @param user The user who made the withdraw
     /// @param amount The withdrawn JUICE amount
-    event JUICEWithdrawn(address indexed user, uint amount);
+    event JUICEWithdrawn(address indexed user, uint256 amount);
 
     /// @notice Emitted when adding to a staked token amount.
     /// @param user The staker
@@ -44,7 +43,13 @@ interface IJuiceStakerActions {
     /// @param sentiment True if this is a long stake.
     /// @param price The token price.
     /// @param unstakedDiff The unstaked JUICE difference (negative when staking)
-    event StakeAdded(address indexed user, address indexed token, bool sentiment, uint price, int unstakedDiff);
+    event StakeAdded(
+        address indexed user,
+        address indexed token,
+        bool sentiment,
+        uint256 price,
+        int256 unstakedDiff
+    );
 
     /// @notice Emitted when unstaking from a token stake.
     /// @param user The staker
@@ -52,14 +57,19 @@ interface IJuiceStakerActions {
     /// @param sentiment True if this is a long stake.
     /// @param price The token price.
     /// @param unstakedDiff The unstaked JUICE difference (positive when unstaking)
-    event StakeRemoved(address indexed user, address indexed token, bool sentiment, uint price, int unstakedDiff);
-
+    event StakeRemoved(
+        address indexed user,
+        address indexed token,
+        bool sentiment,
+        uint256 price,
+        int256 unstakedDiff
+    );
 
     /// @notice Thrown if the StakeData.token is not supported (i.e. couldn't resolve a price feed for it, or it's on the unsafelist).
     error InvalidToken(address token);
 
-    /// @notice Thrown if
-    /// 1) when deposited amount exceeds the balance, or
-    /// 2) when withdrawn amount exceeds the unstaked JUICE balance.
-    error InsufficientJUICE(uint expected, uint actual);
+    /// @notice Thrown when
+    /// 1) deposited amount exceeds the balance, or
+    /// 2) withdrawn amount exceeds the unstaked JUICE balance.
+    error InsufficientJUICE(uint256 expected, uint256 actual);
 }
